% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveBIP.R
\name{generateLineups}
\alias{generateLineups}
\title{Optimize function}
\usage{
generateLineups()
}
\arguments{
\item{Defaults}{to TRUE}
}
\description{
This function generates 10 best lineups based off the Rotowire NBA DFS
predictions for the given day of NBA games
}
\examples{
generateLineups()
}
\keyword{optimize}
