% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapeData.R
\name{scrapeDFS}
\alias{scrapeDFS}
\title{Scrape Function}
\usage{
scrapeDFS()
}
\arguments{
\item{Defaults}{to TRUE}
}
\description{
This function allows the user to scrape the date from RotoWire and
import it nicely into an R data frame
}
\examples{
scrapeDFS()
}
\keyword{scrape}
